requirements.txt
python-docx>=1.1.0
Pillow>=10.0.0
PyYAML>=6.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
rapidfuzz>=3.0.0
